Extract import drop zone binding
