+++ js/views.js
@@ -4,7 +4,7 @@ import { state } from './state.js';
 import { escapeHTML } from './utils.js';
 import { getActiveData, destroyAllCharts, getEffectiveRangeForDate, getLatestValueIndex } from './data.js';
 import { canonicalMetric } from './wearable-adapters.js';
-import { loadPdfImport } from './import-loader.js';
+import { setupDropZone } from './import-drop-zone.js';
 import { createNavigate, getInitialView as getRouterInitialView } from './views-router.js';
 import { createDashboardPageView } from './dashboard-page-view.js';
 import { createLensPageHandlers } from './lens-pages.js';
@@ -151,62 +151,6 @@ export {
   closeModal,
   rememberModalTrigger,
 };
-
-function setupDropZone() {
-  const dropZone = document.getElementById("drop-zone");
-  if (!dropZone || dropZone.dataset.lazyDropZoneBound === 'true') return;
-  dropZone.dataset.lazyDropZoneBound = 'true';
-  dropZone.addEventListener("click", () => {
-    if (window.isImportRunning && window.isImportRunning()) return;
-    document.getElementById('pdf-input')?.click();
-  });
-  dropZone.addEventListener("dragover", e => {
-    e.preventDefault();
-    if (!(window.isImportRunning && window.isImportRunning())) dropZone.classList.add("drag-over");
-  });
-  dropZone.addEventListener("dragleave", e => {
-    e.preventDefault();
-    dropZone.classList.remove("drag-over");
-  });
-  dropZone.addEventListener("drop", async e => {
-    e.preventDefault();
-    dropZone.classList.remove("drag-over");
-    if (window.isImportRunning && window.isImportRunning()) {
-      window.showNotification?.("Import already in progress", "info");
-      return;
-    }
-    const files = Array.from(e.dataTransfer?.files || []);
-    if (files.length === 0) return;
-    let importMod;
-    try {
-      importMod = await loadPdfImport();
-    } catch (err) {
-      window.showNotification?.('Could not load import module - check your connection and try again.', 'error');
-      return;
-    }
-    const { jsonFiles, pdfFiles, imageFiles, dnaFiles, textFiles, unsupportedCount } = await importMod.classifyImportFiles(files);
-    if (unsupportedCount > 0 && jsonFiles.length === 0 && pdfFiles.length === 0 && imageFiles.length === 0 && dnaFiles.length === 0 && textFiles.length === 0) {
-      window.showNotification?.("Unsupported file type. Use PDF, text, image, JSON, or DNA raw data (.txt/.csv).", "error");
-      return;
-    }
-    for (const f of jsonFiles) window.importDataJSON(f);
-    if (dnaFiles.length > 0) {
-      for (const f of dnaFiles) {
-        const header = await f.slice(0, 1500).text();
-        const fmt = window.detectDNAFile ? window.detectDNAFile(header) : null;
-        if ((fmt === 'mtdna' || fmt === '23andme-mito') && window.handleMtDNAFile) await window.handleMtDNAFile(f);
-        else if (fmt === '23andme-y') { window.showNotification?.('Y-chromosome DNA files are not supported', 'info'); }
-        else await window.handleDNAFile(f);
-      }
-    }
-    else if (textFiles.length > 0) { for (const f of textFiles) await importMod.handleTextFile(f); }
-    else if (imageFiles.length > 0) { for (const f of imageFiles) await importMod.handleImageFile(f); }
-    else {
-      if (pdfFiles.length === 1) await importMod.handlePDFFile(pdfFiles[0]);
-      else if (pdfFiles.length > 1) await importMod.handleBatchPDFs(pdfFiles);
-    }
-  });
-}
 
 function markerHasData(m) { return m.values?.some(v => v !== null) ?? false; }
 function setDetailModalShell(...classes) {
